refactor: Redo add select button for mobile
Redo previous implementation of the select button for mobile. New
method is more sure and does not need to deal CSS media queries.
It detects whether the user's device supports the copy to clipboard
method. If support is found the copy button is shown. Otherwise,
the select button is shown.